allow system scene to update PointerLock
