+++ crates/scene_runner/src/update_scene/pointer_lock.rs
@@ -1,15 +1,17 @@
-use bevy::{
-    prelude::*,
-    window::{CursorGrabMode, PrimaryWindow},
-};
+use std::marker::PhantomData;
+
+use bevy::{ecs::system::SystemParam, prelude::*, window::PrimaryWindow};
 use common::{
-    inputs::POINTER_SET,
-    structs::{AppConfig, CursorLocks, PrimaryCamera},
+    inputs::{Action, SystemAction, POINTER_SET},
+    structs::{ActiveDialog, AppConfig, CursorLocked, CursorLocks, PrimaryCamera},
 };
 use input_manager::{InputManager, InputPriority};
 
-use crate::{renderer_context::RendererSceneContext, SceneSets};
-use dcl::interface::CrdtType;
+use crate::{
+    initialize_scene::SuperUserScene, renderer_context::RendererSceneContext,
+    update_world::AddCrdtInterfaceExt, SceneSets,
+};
+use dcl::interface::{ComponentPosition, CrdtType};
 use dcl_component::{
     proto_components::{
         common::Vector3,
@@ -22,6 +24,10 @@ pub struct PointerLockPlugin;
 
 impl Plugin for PointerLockPlugin {
     fn build(&self, app: &mut App) {
+        app.add_crdt_lww_component::<PbPointerLock, PointerLock>(
+            SceneComponentId::POINTER_LOCK,
+            ComponentPosition::RootOnly,
+        );
         app.add_systems(Update, update_pointer_lock.in_set(SceneSets::Input));
     }
 }
@@ -32,19 +38,78 @@ pub struct CumulativePointerDelta {
     pub since: f32,
 }
 
-fn update_pointer_lock(
+#[derive(Component)]
+pub struct PointerLock(PbPointerLock);
+
+impl From<PbPointerLock> for PointerLock {
+    fn from(value: PbPointerLock) -> Self {
+        Self(value)
+    }
+}
+
+#[derive(SystemParam)]
+pub struct CameraInteractionState<'w, 's> {
+    input_manager: InputManager<'w>,
+    state: Local<'s, (ClickState, f32)>,
+    time: Res<'w, Time>,
+    #[system_param(ignore)]
+    _p: PhantomData<&'s ()>,
+}
+
+#[derive(Default, Clone, Copy, PartialEq, Eq, Debug)]
+pub enum ClickState {
+    #[default]
+    None,
+    Clicked,
+    Held,
+    Released,
+}
+
+impl CameraInteractionState<'_, '_> {
+    pub fn update(&mut self, action: Action) -> ClickState {
+        match self.state.0 {
+            ClickState::None | ClickState::Released => {
+                if self.input_manager.just_down(action, InputPriority::None) {
+                    *self.state = (ClickState::Held, self.time.elapsed_seconds());
+                } else {
+                    self.state.0 = ClickState::None;
+                }
+            }
+            ClickState::Held => {
+                if self.input_manager.just_up(action) {
+                    if self.time.elapsed_seconds() - self.state.1 > 0.25 {
+                        self.state.0 = ClickState::Released;
+                    } else {
+                        self.state.0 = ClickState::Clicked;
+                    }
+                }
+            }
+            ClickState::Clicked => self.state.0 = ClickState::Released,
+        }
+
+        self.state.0
+    }
+}
+
+pub fn update_pointer_lock(
     mut commands: Commands,
     mut scenes: Query<(
         Entity,
         &mut RendererSceneContext,
         Option<&mut CumulativePointerDelta>,
+        Option<&SuperUserScene>,
+        Option<Ref<PointerLock>>,
     )>,
     window: Query<&Window, With<PrimaryWindow>>,
     camera: Query<(&Camera, &GlobalTransform), With<PrimaryCamera>>,
     mut prev_coords: Local<Option<Vec2>>,
-    locks: Res<CursorLocks>,
+    mut locks: ResMut<CursorLocks>,
     config: Res<AppConfig>,
-    input_manager: InputManager,
+    mut mb_state: CameraInteractionState,
+    active_dialog: Res<ActiveDialog>,
+    mut cursor_locked: ResMut<CursorLocked>,
+    mut toggle: Local<bool>,
+    mut last_explicit_tick: Local<u32>,
 ) {
     let Ok(window) = window.get_single() else {
         return;
@@ -80,8 +145,47 @@ fn update_pointer_lock(
     };
     *prev_coords = screen_coordinates;
 
+    // Handle mouse input
+    let mut state = mb_state.update(Action::System(SystemAction::CameraLock));
+    let input_manager = &mb_state.input_manager;
+    if state == ClickState::None
+        && input_manager.just_down(SystemAction::Cancel, InputPriority::None)
+        && *toggle
+    {
+        // override
+        state = ClickState::Released;
+        *toggle = false;
+    }
+
+    if state == ClickState::Clicked {
+        *toggle = !*toggle;
+    }
+
+    let mut camera_locked = !active_dialog.in_use() && (state == ClickState::Held || *toggle);
+
+    for (_, context, _, maybe_super, maybe_lock) in scenes.iter_mut() {
+        if maybe_super.is_some()
+            && maybe_lock
+                .as_ref()
+                .is_some_and(|lock| lock.is_changed() || context.tick_number == *last_explicit_tick)
+        {
+            debug!("lock updated by scene");
+            *toggle = maybe_lock.unwrap().0.is_pointer_locked;
+            camera_locked = *toggle;
+            *last_explicit_tick = context.tick_number;
+        }
+    }
+
+    if camera_locked {
+        locks.0.insert("camera");
+        cursor_locked.0 = true;
+    } else {
+        locks.0.remove("camera");
+        cursor_locked.0 = false;
+    }
+
     let pointer_lock = PbPointerLock {
-        is_pointer_locked: window.cursor.grab_mode == CursorGrabMode::Locked,
+        is_pointer_locked: camera_locked,
     };
 
     let frame_delta = input_manager.get_analog(POINTER_SET, InputPriority::Scene);
@@ -90,7 +194,7 @@ fn update_pointer_lock(
         .and_then(|coords| camera.viewport_to_world(camera_position, coords))
         .map(|ray| Vector3::world_vec_from_vec3(&ray.direction));
 
-    for (entity, mut context, maybe_pointer_delta) in scenes.iter_mut() {
+    for (entity, mut context, maybe_pointer_delta, _, _) in scenes.iter_mut() {
         if let Some(mut pointer_delta) = maybe_pointer_delta {
             if context.last_sent == pointer_delta.since {
                 pointer_delta.delta += frame_delta;
